Fix split logic of canonical into url and version in several places
Removed
- org.hl7.fhir.r4b.renderers.DataRenderer#versionFromCanonical
- org.hl7.fhir.r4b.renderers.DataRenderer#systemFromCanonical
- org.hl7.fhir.r5.renderers.DataRenderer#versionFromCanonical
- org.hl7.fhir.r5.renderers.DataRenderer#systemFromCanonical
- org.hl7.fhir.validation.BaseValidator#versionFromCanonical
- org.hl7.fhir.validation.BaseValidator#systemFromCanonical

Replaced all usages of aforementioned methods by the use of
org.hl7.fhir.utilities.CanonicalPair.

Improved CanonicalPair to avoid multiple lookup of the | character. And
added a few convenience methods to the class.

This fixes the issues of #1611 (not replacing all canonical splits in
other places of the code).